improvement(SavedTripEditor): Make component set page title, convert to TypeScript.
